Add validator for each field. This was done to follow Single Responsibility Principle.
